New approach to graph expansion.
Before, when we reduced thunk A to thunk B, we inserted B into the same
computation that orginally had A. This lead to a violation of some
otherwise nice invariants, like "the thunk of a node never changes".

We also decided to do that before we realized that we needed links.

Now, we give B its own computation, and link A's computation to it.